refactor(trust): break strict<->held import cycle + drop ineffectual Protocol stubs (#1510 BH2)
CI CodeQL flagged 6 structural alerts on the BH2 diff: 5 py/unsafe-cyclic-import
errors (the 5 cross-imported names in the strict<->held module-level cycle) + 1
py/ineffectual-statement notice.

Cycle: held.py imported Verdict + EnforcementRecord from strict at module scope
(runtime), while strict imported ExpiryDisposition/HeldActionStore/ReviewerDecision
from held — a bidirectional module-level edge where a name "may not be defined"
depending on import order. Fix: extract the two shared RUNTIME value types
(Verdict, EnforcementRecord) into a new leaf module _types.py that imports neither
strict nor held. Both modules now import these from _types; held no longer imports
strict at module scope at all, so the edge is one-way (strict->held stays
TYPE_CHECKING-only + lazy in-function). strict re-exports Verdict + EnforcementRecord
(module-level import + __all__) so every `from kailash.trust.enforce.strict import
Verdict / EnforcementRecord` caller — including the enforce package __init__ —
resolves unchanged (same object identity). Purely import-graph restructuring, zero
runtime behavior change.

Notice: the HeldActionStore Protocol methods used `docstring + ...`; a bare `...`
expression statement is ineffectual (py/ineffectual-statement). Dropped the `...`
— the docstring is a complete, valid body for a Protocol method (never executed),
and runtime_checkable only inspects method names.

Both import orders (strict-first, held-first) import cleanly; public API paths +
strict's re-export identities verified; pre-commit (black/isort/ruff/type-annos/
Tier-1) green; HITL cluster 72 + broader trust 4837 pass under -W error.
